feat: allow disabling the legacy block prop and artificially reduce block size (#2512)
Closes https://github.com/celestiaorg/celestia-core/issues/2236 and
https://github.com/celestiaorg/celestia-core/issues/2238

I won't backport this PR to main, given main disables by default the
legacy block prop. But I will create another PR that introduces this
change along with the one that added the `disable_block_prop` config.

+++ state/execution_test.go
@@ -802,7 +802,7 @@ func BenchmarkCreateProposalBlock(b *testing.B) {
 	defer cancel()
 
 	state, stateDB, privVals := makeState(1, height)
-	state.ConsensusParams.Block.MaxBytes = types.MaxBlockSizeBytes
+	state.ConsensusParams.Block.MaxBytes = int64(types.MaxBlockSizeBytes)
 	stateStore := sm.NewStore(stateDB, sm.StoreOptions{
 		DiscardABCIResponses: false,
 	})
